Task4 Stage4 FinalVersion

+++ Task4/StatisticApi/Program.cs
@@ -4,25 +4,25 @@ using Dal.Statistic.Repositories.Interfaces;
 using Microsoft.EntityFrameworkCore;
 using StatisticApi.Controllers.mapping;
 using Task4.Statistics;
+using Task4.Statistics.Api;
 using Task4.Statistics.Interfaces;
 
 var builder = WebApplication.CreateBuilder(args);
-
-//путь до файла со статистикой
-var data = builder.Configuration.GetValue<string>("PathToData");
 
 // подключение к бд
 builder.Services.AddDbContext<DataContext>(options =>
 {
     options.UseNpgsql(builder.Configuration.GetConnectionString("DefaultConnection"));
-});
+}, ServiceLifetime.Singleton);
+
 //добавление зависимостей
 builder.Services.AddAutoMapper(typeof(StatisticProfile));
 builder.Services.AddTransient<IConsoleStatisticManager, ConsoleStatisticManager>();
-builder.Services.AddTransient<IStati
-builder.Services.AddScoped<IStatisticKeyRepository, StatisticKeyRepository>();
-builder.Services.AddScoped<IStatisticValueRepository, StatisticValueRepository>();
-builder.Services.AddSingleton<IStatisticData, StatisticsData>();
+builder.Services.AddTransient<IStatisticManager, StatisticManager>();
+builder.Services.AddTransient<IStatisticKeyRepository, StatisticKeyRepository>();
+builder.Services.AddTransient<IStatisticValueRepository, StatisticValueRepository>();
+builder.Services.AddSingleton<Task4.Program>();
+
 
 //добавление контролллеров и маршрутизации
 builder.Services.AddMvc();
@@ -39,15 +39,19 @@ var app = builder.Build();
 app.UseHttpsRedirection();
 app.UseRouting();
 
-// Откючаем (комментируем) если не требуется отчистка бд 
-// т.к. все данные из бд будут удаленны
-using (var scope = 
-       app.Services.CreateScope())
-using (var context = scope.ServiceProvider.GetService<DataContext>())
+// проверка на необходимость обновления базы данных
+//указывается в appsettings.json
+if (bool.Parse(builder.Configuration.GetSection("UpdateDatabase").Value))
 {
-    context.Database.EnsureDeleted();
-    context.Database.EnsureCreated();
+    using (var scope = 
+           app.Services.CreateScope())
+    using (var context = scope.ServiceProvider.GetService<DataContext>())
+    {
+        context.Database.EnsureDeleted();
+        context.Database.EnsureCreated();
+    }
 }
+
 
 app.MapControllerRoute(
     name: "default",
@@ -55,8 +59,8 @@ app.MapControllerRoute(
 
 //инициализация потоков для приложения обработки запросов и конмольного приложения
 var api = new Thread(app.Run);
-var console = new Thread(Task4.Program.Main);
-
+var service = app.Services.GetService<Task4.Program>();
+var console = new Thread(service.ReadConsoleCommand);
 //запуск приложения
 api.Start();
-//console.Start();
+console.Start();
